old/requirements.txt
click
fastapi
uvicorn[standard]
sentence-transformers
chromadb
llama-cpp-python
PyPDF2
pdfminer.six
openai
